These increases to variable size seem sensible. Thanks.

+++ src/nrutil.h
@@ -58,14 +58,14 @@ int *ivector(long nl, long nh);
 char *cvector(long nl, long nh);
 unsigned long *lvector(long nl, long nh);
 double *dvector(long nl, long nh);
-float **matrix(long nrl, long nrh, long ncl, long nch);
+double **matrix(long nrl, long nrh, long ncl, long nch);
 char **cmatrix(long nrl, long nrh, long ncl, long nch);
 double **dmatrix(long nrl, long nrh, long ncl, long nch);
 int **imatrix(long nrl, long nrh, long ncl, long nch);
-float **submatrix(float **a, long oldrl, long oldrh, long oldcl, long oldch,
+double **submatrix(float **a, long oldrl, long oldrh, long oldcl, long oldch,
                   long newrl, long newcl);
-float **convert_matrix(float *a, long nrl, long nrh, long ncl, long nch);
-float ***f3tensor(long nrl, long nrh, long ncl, long nch, long ndl, long ndh);
+double **convert_matrix(float *a, long nrl, long nrh, long ncl, long nch);
+double ***f3tensor(long nrl, long nrh, long ncl, long nch, long ndl, long ndh);
 double ***d3tensor(long nrl, long nrh, long ncl, long nch, long ndl, long ndh);
 char ***c3d(long nrl, long nrh, long ncl, long nch, long ndl, long ndh);
 void free_vector(float *v, long nl, long nh);
@@ -73,13 +73,13 @@ void free_ivector(int *v, long nl, long nh);
 void free_cvector(char *v, long nl, long nh);
 void free_lvector(unsigned long *v, long nl, long nh);
 void free_dvector(double *v, long nl, long nh);
-void free_matrix(float **m, long nrl, long nrh, long ncl, long nch);
+void free_matrix(double **m, long nrl, long nrh, long ncl, long nch);
 void free_cmatrix(char **m, long nrl, long nrh, long ncl, long nch);
 void free_dmatrix(double **m, long nrl, long nrh, long ncl, long nch);
 void free_imatrix(int **m, long nrl, long nrh, long ncl, long nch);
-void free_submatrix(float **b, long nrl, long nrh, long ncl, long nch);
-void free_convert_matrix(float **b, long nrl, long nrh, long ncl, long nch);
-void free_f3tensor(float ***t, long nrl, long nrh, long ncl, long nch,\
+void free_submatrix(double **b, long nrl, long nrh, long ncl, long nch);
+void free_convert_matrix(double **b, long nrl, long nrh, long ncl, long nch);
+void free_f3tensor(double ***t, long nrl, long nrh, long ncl, long nch,\
                    long ndl, long ndh);
 void free_d3tensor(double ***t, long nrl, long nrh, long ncl, long nch,\
                    long ndl, long ndh);
